Implemented OneTimeEmitter and an IO System for it. The Emitter is based on the source code of ConfigurableEmitter, however it will not emmit particles based on time, but on the call of the method spawn. The IO is based on ParticleIO source code and just modified to OneTimeEmmiter
Every platform is now also an emitter. This was done to be able to spawn
"platform particles" when the player hits the platform.

The player now has a fixed hit box instead of using the boundings of the
sprite for collision. By doing this strange situations (e.g. the player
appears to stand in the air) will no longer happen or only very
unlikely.